✅🧱Adding OTP and email checks before email system

+++ back-end/src/users/users.ts
@@ -16,6 +16,18 @@ const path = require('path');
 const multer = require('multer');
 const userpicture = multer();
 const { v4: uuidv4 } = require('uuid');
+const nodemailer = require('nodemailer');
+
+//=============================================================================================//
+//Nodemailer email connection 
+//=============================================================================================//
+var emailer = nodemailer.createTransport({
+  service: 'gmail',
+  auth: {
+    user: process.env.EMAIL,
+    pass: process.env.PASSWORD
+  }
+});
 
 //=============================================================================================//
 //Helper Functions 
@@ -87,8 +99,12 @@ function createID2(length: any) {
 // USER API
 //=============================================================================================//
 const allowedOrigins = [
+  'capacitor://localhost',
+  'ionic://localhost',
+  'http://localhost',
+  'http://localhost:8080',
+  'http://localhost:8100',
   'http://localhost:3000',
-  'http://localhost:8100'
 ];
 const corsOptions = {
   origin: (origin: any, callback: any) => {
@@ -419,9 +435,13 @@ const users = express.Router()
    */
   .post('/users/user', cors(corsOptions), async (req: any, res: any) => {
     try {
-      const result = await userRepository.saveUser(req.body.email,req.body.name,req.body.surname,req.body.number,req.body.username,req.body.password);
-      const results = { 'success': true, 'results': (result) ? result.rows : null};
-      res.json( {results, body: result} );
+      if (req.body.email.toLowerCase().match(/^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/))
+      {
+        const result = await userRepository.saveUser(req.body.email,req.body.name,req.body.surname,req.body.number,req.body.username,req.body.password);
+        res.json({'success':true})
+      } else {
+        res.json({'success':false, 'message':'Invalid email entered!'})
+      }
     } catch (err) {
       const results = { 'success': false, 'results': err };
       console.error(err);
@@ -475,11 +495,17 @@ const users = express.Router()
    .post('/users/user/OTP', cors(corsOptions), async (req: any, res: any) => {
     try {
       const query = req.body;
-      let result = await userOTPRepository.deleteUserOTP(query.email);
-      const newOTP = createID2(6);
-      result = await userOTPRepository.saveUserOTP(query.email,newOTP);
-      const results = { 'success': true };
-      res.json(results);
+      let user = await userRepository.findByEmail(query.email);
+      console.log(user);
+      if(user != null && user.email == query.email)
+      {
+        let result = await userOTPRepository.deleteUserOTP(query.email);
+        const newOTP = createID2(6);
+        result = await userOTPRepository.saveUserOTP(query.email,newOTP);
+        res.json({ 'success': true });
+      } else {
+        res.json({ 'success': false ,'message':'User does not exist!' });
+      }
     } catch (err) {
       const results = { 'success': false, 'results': err };
       console.error(err);
